Split off CLI internals to separate framework (for SPM testing)
